Started building baseline firebase/next.js configuration.
Hey future bobby. this commit is a bit a weird because I'm in drafting mode. This app is getting a bit bigger than expected so I'm just committing at a randomly place in case I want to roll back. I have made most of the firebase connection, but not much of the app at this point.